requirements.txt
numpy
pillow
matplotlib
tqdm
scikit-learn
torch
torchvision
timm
datasets
seaborn
grad-cam
streamlit
opencv-python
pytest
python-dotenv